ass5

+++ 223C/assignment5/find.c
@@ -3,5 +3,5 @@
 
 #include "student.h"
 
-extern struct Student* get(FILE*, long);
+extern struct Student* get(FILE*, long long);
 extern int search(FILE*, unsigned int);
